+++ views/amendment/ajax_diff.php
@@ -13,7 +13,12 @@ echo '<article class="ajaxAmendment">';
 $link = Html::encode(\app\components\UrlHelper::createAmendmentUrl($amendment));
 echo '<a href="' . $link . '" class="amendmentLink"><span class="glyphicon glyphicon-share"></span></a>';
 
-if ($amendment->changeEditorial != '') {
+echo '<section class="initiator">';
+echo '<h4>' . \Yii::t('amend', 'initiator') . '</h4>';
+echo '<div class="paragraph">' . Html::encode($amendment->getInitiatorsStr()) . '</div>';
+echo '</section>';
+
+if ($amendment->changeEditorial) {
     echo '<section id="section_editorial" class="motionTextHolder">';
     echo '<h3 class="green">' . \Yii::t('amend', 'editorial_hint') . '</h3>';
     echo '<div class="paragraph"><div class="text motionTextFormattings">';