+++ History.md
@@ -3,6 +3,7 @@
 ### Version 4.1.1 [not released yet]
 
 - In the list of consultations, the newest consultation now appears at the top.
+- Clicking on changes in the public version of a merging draft now always opens a tooltip with the summary of the amendment, including the proposers of that amendment.
 - Bugfix: the date of the last saved draft when merging motions was not set correctly on Safari.
 
 ## Version 4.1.0 (2018-11-17)